backend/requirements.txt
django>=5.0
djangorestframework>=3.15
drf-spectacular>=0.27
